renamed command folder to scripts and updated help and project index to reflect the rename
