+++ script/templates/bookshelf.js
