+++ script/templates/dumbReduxContainerTemplate.js
