+++ script/templates/enzoBookshelfModelTemplate.js
