+++ script/templates/enzoCreateBookshelfModel.js
