+++ script/templates/enzoCreateMongooseModel.js
