+++ script/templates/enzoDumbComponentTemplate.js
