+++ script/templates/enzoEndpointTemplate.js
