+++ script/templates/migrationTemplate.js
