+++ script/templates/reducerTemplate.js
