+++ script/templates/reduxContainerTemplate.js
