+++ script/templates/schemaTemplate.js
