+++ script/templates/smartComponentTemplate.js
