+++ script/templates/statefulComponent.js
