+++ script/templates/statelessComponent.js
